requirements.txt
haystack-ai>=2.12.0
hayhooks>=0.5.0
lancedb>=0.20.0
sentence-transformers>=3.4.0
pypdf>=5.0.0
python-docx>=1.1.0
odfpy>=1.4.1
watchdog>=6.0.0
markdown-it-py>=3.0.0
mdit-plain>=1.0.0
nltk>=3.9.1
mcp>=1.20.0
pymupdf>=1.25.0
pillow>=11.0.0
